find_first_divergence: parse fractional songlength secs; record Hubbard partials
- find_first_divergence.py crashed on Songlengths.md5 M:SS.mmm entries
  (int('44.554')); parse seconds as float.
- Record project_hubbard_remaining_partials: the 6 still-partial Hubbard
  subtunes after Monty. Human_Race(4)+Battle(1) are CIA-timed with a
  note-onset one-frame timing shift (90% per-play byte-exact; py65 logic is
  correct so it's a first-play/CIA-dispatch subtlety py65 can't model; ruled
  out RAM-init, missing CIA write, global play offset). Devils_Galop(1) is a
  separate vblank bug (dropped V3 freq glide write).

+++ tools/find_first_divergence.py
@@ -89,8 +89,9 @@ def _resolve_duration(orig_path: str, subtune: int,
             if line.startswith(md5):
                 parts = line.rstrip().split('=', 1)[1].split()
                 if subtune < len(parts):
+                    # Songlengths.md5 uses M:SS or M:SS.mmm (fractional secs).
                     m, s = parts[subtune].split(':')
-                    return (int(m) * 60 + int(s)) * 1.1 + 1.0
+                    return (int(m) * 60 + float(s)) * 1.1 + 1.0
                 break
     print(f'warning: subtune {subtune} not in Songlengths.md5 — '
           f'using 60s default', file=sys.stderr)
